pyproject.toml
[project]
name = "multi-agent-a2a"
version = "0.1.0"
description = "Multi-agent AI system with A2A protocol communication"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "a2a-sdk[http-server]>=0.2.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

orchestrator = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
]

research = [
    "crewai>=0.86.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "firecrawl-py>=1.0.0",
    "mcp>=1.0.0",
]

explainer = [
    "pydantic-ai>=0.0.24",
]

knowledge = [
    "agno>=1.0.0",
    "lancedb>=0.15.0",
]

azure = [
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "langchain-openai>=0.2.0",
]

all = [
    "multi-agent-a2a[orchestrator,research,explainer,knowledge,azure,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "shared"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
